requirements.txt
numpy>=1.21.0
pandas>=1.3.0
pyyaml>=6.0
pytest>=7.0.0
jupyter>=1.0.0
scikit-learn>=1.0.0
tqdm>=4.62.0
python-dotenv>=0.19.0
transformers>=4.30.0
torch>=2.0.0
openai>=1.0.0
datasets>=2.12.0 